Add support for custom tmpfs
Sometimes a tmpfs may already be mounted elsewhere and the user
would like to reuse it. For this we now allow a env var to point
us to an existing tmpfs which we use for the debootstrap process.

Also riding on this patch - cosmetic changes to the warn/info/notice
funcs that print msgs to screen.