Centralize and clean up logic that creates partitions
Now we iterate over a list of partitions and create each one
instead of treating the some partitions separately. In order
to support this, I've added more flags to the partition structs
to indicate whether they are 'os' or 'boot' partitions.

The func to create partitions now resides in the base class and
expects a new func to be implemented by derivative classes (which
will return the partition layout to be created).