Make 'partition_layout' file an arg
So that the user can provide the layout in a JSON file and the
different platform specific builders can be reused without needing
the user to write a custom derived class to specify custom layout.

This is plumbed via an 'ENV' var (on the cmdline for Rake tasks).

Also add various layout files (previously defined as vars in the
platform specific disk builder files).