Derived disk builders can specify additional partitions
So that the partition layout files don't have to contain this
knowledge (which the use may be oblivious of). This patch also
changes the various PART_LAYOUT_xxx files to reflect this new
scheme.

Additionally, validate the partition layout by checking for the
presence of exactly one grub cfg partition and at least one OS
partition.