Pass additional options to grub-mkrescue to handle >4GiB rootfs
When creating ISO's, the iso rootfs may be very large (depending on
what packages are installed in it, or if some large payloads are
copied into it). In such cases the resulting squashfs (created
from the rootfs) can be larger than 4GiB. In these cases we need
to pass an additional flag to grub-mkrescue (which gets passed on
to xorriso) to enable newer compatibility mode.

Unfortunately the rootfs is compressed into a single squashfs
since that is what live-boot knows how to boot from. So splitting
individual files inside the rootfs to be <4GiB isn't an option
since everything adds up into a single squashfs file (which has
to be <4GiB). So this is the best path forward.