Make rootfs compression configurable
Earlier the rootfs was always built to be a tar.gz file, however
in some cases we might want to let the user configure this (e.g.
where the next step runs immediately and has to decompress the
tarball to use it).

This patch allows the compression to be configured and also changes
the default setting to be uncompressed (along with Rakefile changes
to expec the new file extension). The disk builder is also changed
so that it can detect the appropriate compression used.