Ensure resulting output file is owned by current user
At the end of a debootstrap run, the resulting rootfs should be
owned by the current user. This may not be the case after 'tar'
completes since it is executed under sudo, so this patch adds
an explicit call to 'chown' to change the ownership.

Also riding on this patch:
- ensure keys in the dict are strings (not symbols)
- fix bad reference to var