Add ability to specify "open ended" partitions/LVs
By specifying an empty 'size_mb', diskbot now allows you to create
partitions (and LVs) that use up the remaining space of the disk or
VG as the case may be (using one of 100% or 100%FREE). This is useful
in cases where we deal with actual devices that we'd like to consume
without needing to know the size of a priori. In such cases we loosen
the disk size validation (since we cannot determine whether an open
ended partition is sufficiently big to hold its LVs if necessary) so
this may result in more runtime errors but the caller is assumed to
have performed these checks himself/herself.